comm - Refactor imports in vector store unit tests and add lazy loading utility
