+++ pkgs/community/swarmauri_community/llms/concrete/__init__.py
@@ -1,4 +1,12 @@
-from swarmauri_community.llms.concrete.LeptonAIImgGenModel import LeptonAIImgGenModel
-from swarmauri_community.llms.concrete.LeptonAIModel import LeptonAIModel
+from swarmauri.utils._lazy_import import _lazy_import
 
-__all__ = ["LeptonAIImgGenModel", "LeptonAIModel"]
+llms_files = [
+    ("swarmauri_community.llms.concrete.LeptonAIImgGenModel", "LeptonAIImgGenModel"),
+    ("swarmauri_community.llms.concrete.LeptonAIModel", "LeptonAIModel"),
+    ("swarmauri_community.llms.concrete.PytesseractImg2TextModel", "PytesseractImg2TextModel"),
+]
+
+for module_name, class_name in llms_files:
+    globals()[class_name] = _lazy_import(module_name, class_name)
+
+__all__ = [class_name for _, class_name in llms_files]
