+++ pkgs/community/swarmauri_community/retrievers/concrete/__init__.py
@@ -1,5 +1,10 @@
-# -*- coding: utf-8 -*-
+from swarmauri.utils._lazy_import import _lazy_import
 
-from swarmauri_community.retrievers.concrete.RedisDocumentRetriever import (
-    RedisDocumentRetriever,
-)
+retriever_files = [
+    ("swarmauri_community.retrievers.concrete.RedisDocumentRetriever", "RedisDocumentRetriever"),
+]
+
+for module_name, class_name in retriever_files:
+    globals()[class_name] = _lazy_import(module_name, class_name)
+
+__all__ = [class_name for _, class_name in retriever_files]
