+++ pkgs/community/swarmauri_community/tools/concrete/__init__.py
@@ -1,31 +1,33 @@
-from swarmauri_community.tools.concrete.CaptchaGeneratorTool import CaptchaGeneratorTool
-from swarmauri_community.tools.concrete.DaleChallReadabilityTool import (
-    DaleChallReadabilityTool,
-)
-from swarmauri_community.tools.concrete.DownloadPdfTool import DownloadPDFTool
-from swarmauri_community.tools.concrete.EntityRecognitionTool import (
-    EntityRecognitionTool,
-)
-from swarmauri_community.tools.concrete.FoliumTool import FoliumTool
-from swarmauri_community.tools.concrete.GithubBranchTool import GithubBranchTool
-from swarmauri_community.tools.concrete.GithubCommitTool import GithubCommitTool
-from swarmauri_community.tools.concrete.GithubIssueTool import GithubIssueTool
-from swarmauri_community.tools.concrete.GithubPRTool import GithubPRTool
-from swarmauri_community.tools.concrete.GithubRepoTool import GithubRepoTool
-from swarmauri_community.tools.concrete.GithubTool import GithubTool
-from swarmauri_community.tools.concrete.GmailReadTool import GmailReadTool
-from swarmauri_community.tools.concrete.GmailSendTool import GmailSendTool
-from swarmauri_community.tools.concrete.LexicalDensityTool import LexicalDensityTool
-from swarmauri_community.tools.concrete.PsutilTool import PsutilTool
-from swarmauri_community.tools.concrete.QrCodeGeneratorTool import QrCodeGeneratorTool
-from swarmauri_community.tools.concrete.SentenceComplexityTool import (
-    SentenceComplexityTool,
-)
-from swarmauri_community.tools.concrete.SentimentAnalysisTool import (
-    SentimentAnalysisTool,
-)
-from swarmauri_community.tools.concrete.SMOGIndexTool import SMOGIndexTool
-from swarmauri_community.tools.concrete.WebScrapingTool import WebScrapingTool
-from swarmauri_community.tools.concrete.ZapierHookTool import ZapierHookTool
+from swarmauri.utils._lazy_import import _lazy_import
 
-# from swarmauri_community.tools.concrete.PaCMAPTool import PaCMAPTool
+tool_files = [
+    ("swarmauri_community.tools.concrete.CaptchaGeneratorTool", "CaptchaGeneratorTool"),
+    ("swarmauri_community.tools.concrete.DaleChallReadabilityTool", "DaleChallReadabilityTool"),
+    ("swarmauri_community.tools.concrete.DownloadPdfTool", "DownloadPDFTool"),
+    ("swarmauri_community.tools.concrete.EntityRecognitionTool", "EntityRecognitionTool"),
+    ("swarmauri_community.tools.concrete.FoliumTool", "FoliumTool"),
+    ("swarmauri_community.tools.concrete.GithubBranchTool", "GithubBranchTool"),
+    ("swarmauri_community.tools.concrete.GithubCommitTool", "GithubCommitTool"),
+    ("swarmauri_community.tools.concrete.GithubIssueTool", "GithubIssueTool"),
+    ("swarmauri_community.tools.concrete.GithubPRTool", "GithubPRTool"),
+    ("swarmauri_community.tools.concrete.GithubRepoTool", "GithubRepoTool"),
+    ("swarmauri_community.tools.concrete.GithubTool", "GithubTool"),
+    ("swarmauri_community.tools.concrete.GmailReadTool", "GmailReadTool"),
+    ("swarmauri_community.tools.concrete.GmailSendTool", "GmailSendTool"),
+    ("swarmauri_community.tools.concrete.LexicalDensityTool", "LexicalDensityTool"),
+    ("swarmauri_community.tools.concrete.PsutilTool", "PsutilTool"),
+    ("swarmauri_community.tools.concrete.QrCodeGeneratorTool", "QrCodeGeneratorTool"),
+    ("swarmauri_community.tools.concrete.SentenceComplexityTool", "SentenceComplexityTool"),
+    ("swarmauri_community.tools.concrete.SentimentAnalysisTool", "SentimentAnalysisTool"),
+    ("swarmauri_community.tools.concrete.SMOGIndexTool", "SMOGIndexTool"),
+    ("swarmauri_community.tools.concrete.WebScrapingTool", "WebScrapingTool"),
+    ("swarmauri_community.tools.concrete.ZapierHookTool", "ZapierHookTool"),
+    # ("swarmauri_community.tools.concrete.PaCMAPTool", "PaCMAPTool"),
+]
+
+# Lazy loading of tools, storing them in variables
+for module_name, class_name in tool_files:
+    globals()[class_name] = _lazy_import(module_name, class_name)
+
+# Adding the lazy-loaded tools to __all__
+__all__ = [class_name for _, class_name in tool_files]
