+++ pkgs/community/swarmauri_community/vector_stores/concrete/Doc2VecVectorStore.py
