+++ pkgs/community/tests/unit/vector_stores/CloudWeaviateVectorStore_test.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 from swarmauri.documents.concrete.Document import Document
-from swarmauri_community.vector_stores.CloudWeaviateVectorStore import (
+from swarmauri_community.vector_stores.concrete.CloudWeaviateVectorStore import (
     CloudWeaviateVectorStore,
 )
 from dotenv import load_dotenv
