+++ pkgs/community/tests/unit/vector_stores/PersistentQdrantVectorStore_test.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 from swarmauri.documents.concrete.Document import Document
-from swarmauri_community.vector_stores.PersistentQdrantVectorStore import (
+from swarmauri_community.vector_stores.concrete.PersistentQdrantVectorStore import (
     PersistentQdrantVectorStore,
 )
