+++ pkgs/community/tests/unit/vector_stores/RedisVectorStore_test.py
@@ -1,8 +1,6 @@
 import pytest
-import numpy as np
 from swarmauri.documents.concrete.Document import Document
-from swarmauri_community.vector_stores.RedisVectorStore import RedisVectorStore
-from swarmauri.documents.concrete.Document import Document
+from swarmauri_community.vector_stores.concrete.RedisVectorStore import RedisVectorStore
 from dotenv import load_dotenv
 from os import getenv
